Implement collapsible academy navigation bar (#1595)
* Implemented collapsible desktop Academy Navbar

* Consolidated all navbar related files and tests under directory /src/commons/navigationBar

* Updated to relative imports
