+++ src/commons/navigationBar/NavigationBar.tsx
@@ -6,7 +6,9 @@ import {
   Navbar,
   NavbarDivider,
   NavbarGroup,
-  NavbarHeading
+  NavbarHeading,
+  Position,
+  Tooltip
 } from '@blueprintjs/core';
 import { IconNames } from '@blueprintjs/icons';
 import classNames from 'classnames';
@@ -17,7 +19,8 @@ import { NavLink } from 'react-router-dom';
 import { Role } from '../application/ApplicationTypes';
 import Dropdown from '../dropdown/Dropdown';
 import Constants from '../utils/Constants';
-import NavigationBarMobileSideMenu from './NavigationBarMobileSideMenu';
+import AcademyNavigationBar from './subcomponents/AcademyNavigationBar';
+import NavigationBarMobileSideMenu from './subcomponents/NavigationBarMobileSideMenu';
 
 type NavigationBarProps = DispatchProps & StateProps;
 
@@ -33,6 +36,7 @@ type StateProps = {
 
 const NavigationBar: React.FC<NavigationBarProps> = props => {
   const [mobileSideMenuOpen, setMobileSideMenuOpen] = React.useState(false);
+  const [desktopMenuOpen, setDesktopMenuOpen] = React.useState(true);
   const isMobile = useMediaQuery({ maxWidth: 768 });
 
   const playgroundOnlyNavbarLeft = (
@@ -126,6 +130,20 @@ const NavigationBar: React.FC<NavigationBarProps> = props => {
         <div className="navbar-button-text hidden-sm hidden-xs">Contributors</div>
       </NavLink>
 
+      {!Constants.playgroundOnly && props.role && !isMobile && (
+        <>
+          <NavbarDivider className="default-divider" />
+          <Tooltip content="Toggle Telebay" position={Position.BOTTOM}>
+            <Button
+              onClick={() => setDesktopMenuOpen(!desktopMenuOpen)}
+              icon={IconNames.COMPASS}
+              minimal={true}
+              style={{ outline: 'none' }}
+            />
+          </Tooltip>
+        </>
+      )}
+
       <div className="visible-xs">
         <NavbarDivider className="thin-divider" />
       </div>
@@ -147,6 +165,10 @@ const NavigationBar: React.FC<NavigationBarProps> = props => {
           : desktopNavbarLeft}
         {commonNavbarRight}
       </Navbar>
+
+      {!Constants.playgroundOnly && props.role && !isMobile && desktopMenuOpen && (
+        <AcademyNavigationBar role={props.role} />
+      )}
     </>
   );
 };
