+++ src/commons/navigationBar/__tests__/NavigationBar.tsx
@@ -1,9 +1,8 @@
 import { shallow } from 'enzyme';
 import * as React from 'react';
-import { Role } from 'src/commons/application/ApplicationTypes';
 
+import { Role } from '../../application/ApplicationTypes';
 import NavigationBar from '../NavigationBar';
-import NavigationBarMobileSideMenu from '../NavigationBarMobileSideMenu';
 
 test('NavigationBar renders "Not logged in" correctly', () => {
   const props = {
@@ -21,24 +20,5 @@ test('NavigationBar renders correctly with student role', () => {
     role: Role.Student
   };
   const tree = shallow(<NavigationBar {...props} />);
-  expect(tree.debug()).toMatchSnapshot();
-});
-
-test('NavigationBarMobileSideMenu renders "Not logged in" correctly', () => {
-  const props = {
-    isOpen: true,
-    onClose: () => {}
-  };
-  const tree = shallow(<NavigationBarMobileSideMenu {...props} />);
-  expect(tree.debug()).toMatchSnapshot();
-});
-
-test('NavigationBarMobileSideMenu renders correctly with student role', () => {
-  const props = {
-    isOpen: true,
-    onClose: () => {},
-    role: Role.Student
-  };
-  const tree = shallow(<NavigationBarMobileSideMenu {...props} />);
   expect(tree.debug()).toMatchSnapshot();
 });
