+++ src/commons/navigationBar/subcomponents/__tests__/NavigationBarMobileSideMenu.tsx
@@ -1,0 +1,24 @@
+import { shallow } from 'enzyme';
+import * as React from 'react';
+
+import { Role } from '../../../application/ApplicationTypes';
+import NavigationBarMobileSideMenu from '../NavigationBarMobileSideMenu';
+
+test('NavigationBarMobileSideMenu renders "Not logged in" correctly', () => {
+  const props = {
+    isOpen: true,
+    onClose: () => {}
+  };
+  const tree = shallow(<NavigationBarMobileSideMenu {...props} />);
+  expect(tree.debug()).toMatchSnapshot();
+});
+
+test('NavigationBarMobileSideMenu renders correctly with student role', () => {
+  const props = {
+    isOpen: true,
+    onClose: () => {},
+    role: Role.Student
+  };
+  const tree = shallow(<NavigationBarMobileSideMenu {...props} />);
+  expect(tree.debug()).toMatchSnapshot();
+});
